core: move blobTxForPool type to blobpool.go
